Add main.rb as responsible for user interaction
